Clean up some warnings.

+++ src/wildcard.cpp
@@ -45,7 +45,7 @@ void wildcard_expand(const std::shared_ptr<char> arg)
   Command::currentCommand.wc_collector.clear();
   Command::currentCommand.wc_collector.shrink_to_fit();
 
-  for (int i = 0; i < results.gl_pathc; i++) {
+  for (size_t i = 0; i < results.gl_pathc; i++) {
     Command::currentCommand.wc_collector.push_back(results.gl_pathv[i]);
   }
   globfree(&results);
